Improving Services Performances  (#65)
